requirements.txt
aniso8601==9.0.1
anyio==3.6.1
autopep8==1.7.0
backports.cached-property==1.0.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
fastapi==0.79.0
graphene==3.1
graphql-core==3.2.1
graphql-relay==3.2.0
h11==0.13.0
idna==3.3
pycodestyle==2.9.1
pydantic==1.9.1
Pygments==2.12.0
python-dateutil==2.8.2
python-multipart==0.0.5
requests==2.28.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
starlette-graphene3==0.6.0
strawberry-graphql==0.124.0
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
